# openapi/legal-rag-openapi.yaml
openapi: 3.1.0
info:
  title: Legal Agentic RAG Internal Platform API
  version: 0.1.0
  description: Spec-first API for corpus ingestion, QA runtime, evaluation, gold annotation,
    and synthetic dataset generation.
servers:
- url: https://api.example.internal
...
paths:
...
  /v1/gold/datasets:
    post:
      tags:
      - Gold
      summary: Create a gold dataset
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldDataset'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                project_id:
                  type: string
                  format: uuid
                name:
                  type: string
                version:
                  type: string
                base_dataset_id:
                  type:
                  - string
                  - 'null'
                  format: uuid
              required:
              - project_id
              - name
              - version
  /v1/gold/datasets/{goldDatasetId}:
    get:
      tags:
      - Gold
      summary: Get one gold dataset
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldDataset'
      parameters:
      - name: goldDatasetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold dataset UUID
  /v1/gold/datasets/{goldDatasetId}/questions:
    post:
      tags:
      - Gold
      summary: Create or import a gold question
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldQuestion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoldQuestion'
      parameters:
      - name: goldDatasetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold dataset UUID
  /v1/gold/questions/{goldQuestionId}:
    patch:
      tags:
      - Gold
      summary: Update one gold question
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldQuestion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoldQuestion'
      parameters:
      - name: goldQuestionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold question UUID
  /v1/gold/questions/{goldQuestionId}/source-sets:
    post:
      tags:
      - Gold
      summary: Add an alternative sufficient source set
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  source_set_id:
                    type: string
                    format: uuid
                required:
                - source_set_id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                is_primary:
                  type: boolean
                page_ids:
                  type: array
                  items:
                    type: string
                  uniqueItems: true
                notes:
                  type:
                  - string
                  - 'null'
              required:
              - is_primary
              - page_ids
      parameters:
      - name: goldQuestionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold question UUID
  /v1/gold/questions/{goldQuestionId}/review:
    post:
      tags:
      - Gold
      summary: Submit review decision for a gold question
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                - status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                decision:
                  type: string
                  enum:
                  - approve
                  - changes_requested
                  - lock
                comment:
                  type:
                  - string
                  - 'null'
              required:
              - decision
      parameters:
      - name: goldQuestionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold question UUID
  /v1/gold/datasets/{goldDatasetId}/lock:
    post:
      tags:
      - Gold
      summary: Lock a gold dataset version
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                - status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason:
                  type:
                  - string
                  - 'null'
      parameters:
      - name: goldDatasetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold dataset UUID
  /v1/gold/datasets/{goldDatasetId}/export:
    get:
      tags:
      - Gold
      summary: Export gold dataset JSON
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  artifact_url:
                    type: string
                    format: uri
                required:
                - artifact_url
      parameters:
      - name: goldDatasetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Gold dataset UUID
...
components:
  schemas:
...
    GoldQuestion:
      title: GoldQuestion
      description: Reviewed gold annotation entry with canonical answer and one or
        more sufficient source sets.
      type: object
      additionalProperties: false
      properties:
        gold_question_id:
          type: string
          format: uuid
        question_id:
          type: string
        gold_dataset_id:
          type: string
          format: uuid
        canonical_answer:
          type:
          - string
          - number
          - boolean
          - 'null'
          - array
        acceptable_answers:
          type: array
          items:
            type:
            - string
            - number
            - boolean
            - 'null'
            - array
        answer_type:
          type: string
          enum:
          - boolean
          - number
          - date
          - name
          - names
          - free_text
        source_sets:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              source_set_id:
                type: string
                format: uuid
              is_primary:
                type: boolean
              page_ids:
                type: array
                items:
                  type: string
                minItems: 0
                uniqueItems: true
              notes:
                type:
                - string
                - 'null'
            required:
            - source_set_id
            - is_primary
            - page_ids
          minItems: 1
        review_status:
          type: string
          enum:
          - draft
          - reviewed
          - locked
        reviewers:
          type: array
          items:
            type: string
        notes:
          type:
          - string
          - 'null'
      required:
      - gold_question_id
      - question_id
      - gold_dataset_id
      - canonical_answer
      - answer_type
      - source_sets
      - review_status
    GoldDataset:
      title: GoldDataset
      description: Versioned gold dataset definition.
      type: object
      additionalProperties: false
      properties:
        gold_dataset_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        status:
          type: string
          enum:
          - draft
          - review
          - locked
          - archived
        base_dataset_id:
          type:
          - string
          - 'null'
          format: uuid
        question_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - gold_dataset_id
      - project_id
      - name
      - version
      - status
      - question_count
      - created_at
      - updated_at
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []